+++ lib/change_health/models/claim/submission/address.rb
@@ -2,11 +2,11 @@ module ChangeHealth
   module Models
     module Claim
       class Address < Model
-        property :address1, required: false
-        property :address2, required: false
-        property :city, required: false
-        property :postalCode, from: :postal_code, required: false
-        property :state, required: false
+        property :address1
+        property :address2
+        property :city
+        property :postalCode, from: :postal_code
+        property :state
       end
     end
   end