+++ lib/change_health/models/claim/submission/dependent.rb
@@ -2,13 +2,13 @@ module ChangeHealth
   module Models
     module Claim
       class Dependent < Model
-        property :address, required: false
-        property :dateOfBirth, from: :date_of_birth, required: false
-        property :firstName, from: :first_name, required: false
-        property :gender, required: false
-        property :lastName, from: :last_name, required: false
-        property :memberId, from: :member_id, required: false
-        property :relationshipToSubscriberCode, from: :relationship_to_subscriber_code, required: false
+        property :address
+        property :dateOfBirth, from: :date_of_birth
+        property :firstName, from: :first_name
+        property :gender
+        property :lastName, from: :last_name
+        property :memberId, from: :member_id
+        property :relationshipToSubscriberCode, from: :relationship_to_subscriber_code
       end
     end
   end