+++ lib/change_health/models/claim/submission/drug_identification.rb
@@ -2,10 +2,10 @@ module ChangeHealth
   module Models
     module Claim
       class DrugIdentification < Model
-        property :measurementUnitCode, from: :measurement_unit_code, required: false
-        property :nationalDrugCode, from: :national_drug_code, required: false
-        property :nationalDrugUnitCount, from: :national_drug_unit_count, required: false
-        property :serviceIdQualifier, from: :service_id_qualifier, required: false
+        property :measurementUnitCode, from: :measurement_unit_code
+        property :nationalDrugCode, from: :national_drug_code
+        property :nationalDrugUnitCount, from: :national_drug_unit_count
+        property :serviceIdQualifier, from: :service_id_qualifier
       end
     end
   end