+++ lib/change_health/models/claim/submission/provider.rb
@@ -1,17 +1,17 @@
 module ChangeHealth
   module Models
     module Claim
-      class Provider < Hashie::Trash
-        property :address, required: false
-        property :contactInformation, from: :contact_information, required: false
-        property :employerId, from: :employer_id, required: false # or ssn
-        property :firstName, from: :first_name, required: false
-        property :lastName, from: :last_name, required: false
-        property :organizationName, from: :organization_name, required: false
-        property :npi, required: false
-        property :providerType, from: :provider_type, required: false
-        property :taxonomyCode, from: :taxonomy_code, required: false
-        property :ssn, required: false # or employer id
+      class Provider < Model
+        property :address
+        property :contactInformation, from: :contact_information
+        property :employerId, from: :employer_id # or ssn
+        property :firstName, from: :first_name
+        property :lastName, from: :last_name
+        property :npi
+        property :organizationName, from: :organization_name
+        property :providerType, from: :provider_type
+        property :ssn # or employer id
+        property :taxonomyCode, from: :taxonomy_code
       end
     end
   end