+++ test/change_health/models/claim/submission/claim_information_test.rb
@@ -2,14 +2,17 @@ require 'test_helper'
 
 class ClaimInformationTest < Minitest::Test
   describe 'claim information' do
-    let(:claim_information) { ChangeHealth::Models::Claim::ClaimInformation.new(benefits_assignment_certification_indicator: 'cat') }
+    let(:claim_information) do
+      ChangeHealth::Models::Claim::ClaimInformation.new(benefits_assignment_certification_indicator: 'cat')
+    end
 
     describe 'object' do
       describe 'serializes' do
         it 'can serialize to json' do
           result = JSON.parse(claim_information.to_json)
 
-          assert_equal(claim_information.benefitsAssignmentCertificationIndicator, result['benefitsAssignmentCertificationIndicator'])
+          assert_equal(claim_information.benefitsAssignmentCertificationIndicator,
+                       result['benefitsAssignmentCertificationIndicator'])
         end
       end
 
@@ -20,7 +23,7 @@ class ClaimInformationTest < Minitest::Test
 
         it 'can add a service line' do
           serviceLine = {
-            professionalService: "cat"
+            professionalService: 'cat'
           }
           claim_information.add_service_line(serviceLine)
           assert_equal(1, claim_information.serviceLines.size)
@@ -33,12 +36,13 @@ class ClaimInformationTest < Minitest::Test
 
         it 'can add a health_care_code_information' do
           health_care_code_information = {
-            "diagnosisTypeCode": "ABK",
-            "diagnosisCode": "S93401A"
+            "diagnosisTypeCode": 'ABK',
+            "diagnosisCode": 'S93401A'
           }
           claim_information.add_health_care_code_information(health_care_code_information)
           assert_equal(1, claim_information.healthCareCodeInformation.size)
-          assert_equal(health_care_code_information[:diagnosisTypeCode], claim_information.healthCareCodeInformation.first[:diagnosisTypeCode])
+          assert_equal(health_care_code_information[:diagnosisTypeCode],
+                       claim_information.healthCareCodeInformation.first[:diagnosisTypeCode])
         end
       end
     end